Refactor: move `IngestionStatistics` to own file
