+++ crates/store/re_entity_db/src/entity_db.rs
@@ -1,12 +1,11 @@
 use std::sync::Arc;
 
 use nohash_hasher::IntMap;
-use parking_lot::Mutex;
 
 use re_chunk::{Chunk, ChunkResult, LatestAtQuery, RowId, TimeInt, Timeline, TimelineName};
 use re_chunk_store::{
     ChunkStore, ChunkStoreChunkStats, ChunkStoreConfig, ChunkStoreDiffKind, ChunkStoreEvent,
-    ChunkStoreHandle, ChunkStoreSubscriber, GarbageCollectionOptions, GarbageCollectionTarget,
+    ChunkStoreHandle, ChunkStoreSubscriber as _, GarbageCollectionOptions, GarbageCollectionTarget,
 };
 use re_log_types::{
     ApplicationId, EntityPath, EntityPathHash, LogMsg, ResolvedTimeRange, ResolvedTimeRangeF,
@@ -18,7 +17,7 @@ use re_query::{
 };
 use re_smart_channel::SmartChannelSource;
 
-use crate::{Error, TimesPerTimeline};
+use crate::{Error, TimesPerTimeline, ingestion_statistics::IngestionStatistics};
 
 // ----------------------------------------------------------------------------
 
@@ -867,86 +866,5 @@ impl re_byte_size::SizeBytes for EntityDb {
             .stats()
             .total()
             .total_size_bytes
-    }
-}
-
-// ----------------------------------------------------------------------------
-
-pub struct IngestionStatistics {
-    store_id: StoreId,
-    e2e_latency_sec_history: Mutex<emath::History<f32>>,
-}
-
-impl ChunkStoreSubscriber for IngestionStatistics {
-    #[inline]
-    fn name(&self) -> String {
-        "rerun.testing.store_subscribers.IngestionStatistics".into()
-    }
-
-    #[inline]
-    fn as_any(&self) -> &dyn std::any::Any {
-        self
-    }
-
-    #[inline]
-    fn as_any_mut(&mut self) -> &mut dyn std::any::Any {
-        self
-    }
-
-    #[inline]
-    fn on_events(&mut self, events: &[ChunkStoreEvent]) {
-        for event in events {
-            if event.store_id == self.store_id {
-                for row_id in event.diff.chunk.row_ids() {
-                    self.on_new_row_id(row_id);
-                }
-            }
-        }
-    }
-}
-
-impl IngestionStatistics {
-    pub fn new(store_id: StoreId) -> Self {
-        let min_samples = 0; // 0: we stop displaying e2e latency if input stops
-        let max_samples = 1024; // don't waste too much memory on this - we just need enough to get a good average
-        let max_age = 1.0; // don't keep too long of a rolling average, or the stats get outdated.
-        Self {
-            store_id,
-            e2e_latency_sec_history: Mutex::new(emath::History::new(
-                min_samples..max_samples,
-                max_age,
-            )),
-        }
-    }
-
-    fn on_new_row_id(&self, row_id: RowId) {
-        if let Ok(duration_since_epoch) = web_time::SystemTime::UNIX_EPOCH.elapsed() {
-            let nanos_since_epoch = duration_since_epoch.as_nanos() as u64;
-
-            // This only makes sense if the clocks are very good, i.e. if the recording was on the same machine!
-            if let Some(nanos_since_log) = nanos_since_epoch.checked_sub(row_id.nanos_since_epoch())
-            {
-                let now = nanos_since_epoch as f64 / 1e9;
-                let sec_since_log = nanos_since_log as f32 / 1e9;
-
-                self.e2e_latency_sec_history.lock().add(now, sec_since_log);
-            }
-        }
-    }
-
-    /// What is the mean latency between the time data was logged in the SDK and the time it was ingested?
-    ///
-    /// This is based on the clocks of the viewer and the SDK being in sync,
-    /// so if the recording was done on another machine, this is likely very inaccurate.
-    pub fn current_e2e_latency_sec(&self) -> Option<f32> {
-        let mut e2e_latency_sec_history = self.e2e_latency_sec_history.lock();
-
-        if let Ok(duration_since_epoch) = web_time::SystemTime::UNIX_EPOCH.elapsed() {
-            let nanos_since_epoch = duration_since_epoch.as_nanos() as u64;
-            let now = nanos_since_epoch as f64 / 1e9;
-            e2e_latency_sec_history.flush(now); // make sure the average is up-to-date.
-        }
-
-        e2e_latency_sec_history.average()
     }
 }
